v0.1.7 beta
Improve the centroid estimation of spike waveform. Template center channel is now defined as the channel where peak power is found. Previously, it was simply the middle channel.
Handles the retrieval of spike waveforms that are out of channel range.
Increase the max number of clusters to select and plot by changing MTracer.ClusteringVM.maxClusOp to 8.